Update
Updated index and style, added <hr> tags and <strong> to bold the instruction titles.  I stylized the site a bit further than I wanted to tonight, mobile view is almost finished, just have to adjust the nutritional values at the bottom of the page...They need to align to the first letter/character.