Update
I updated quite a bit, had to fix the alt tags for the images. Since I wanted the border to show around the image in desktop but not in mobile, I had to set the image twice, but hide/make visible each one what the media query called upon it. I also amended the width of the main card, applied border radius, and adjusted font sizes for everything. I also realized that I needed to change the color of the markers in the ul/ol lists. This should be closer than the 1st submission.